Update SM4
Simple SM4 cipher implementation in `sm4.c`
